+++ inst/doc/etwfe.R
@@ -1,41 +1,83 @@
-## ----include = FALSE----------------------------------------------------------
+## ----include = FALSE------------------------------------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
+options(width = 100)
 options(rmarkdown.html_vignette.check_title = FALSE)
-options(modelsummary_factory_default = "gt")
+options(marginaleffects_safe = FALSE)
+modelsummary::config_modelsummary(startup_message = FALSE)
 
 fixest::setFixest_notes(FALSE)
 
-## -----------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
 # install.packages("did")
 data("mpdta", package = "did")
 head(mpdta)
 
-## -----------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
 library(etwfe)
 
-mod =
-  etwfe(
-    fml  = lemp ~ lpop, # outcome ~ controls
-    tvar = year,        # time variable
-    gvar = first.treat, # group variable
-    data = mpdta,       # dataset
-    vcov = ~countyreal  # vcov adjustment (here: clustered)
-    )
+mod = etwfe(
+  fml  = lemp ~ lpop, # outcome ~ controls
+  tvar = year,        # time variable
+  gvar = first.treat, # group variable
+  data = mpdta,       # dataset
+  vcov = ~countyreal  # vcov adjustment (here: clustered)
+)
 
-## -----------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
 mod
 
-## -----------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
 emfx(mod)
 
-## -----------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
 mod_es = emfx(mod, type = "event")
 mod_es
 
-## -----------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
+plot(mod_es)
+
+## -------------------------------------------------------------------------------------------------
+mod_es2 = etwfe(
+  lemp ~ lpop, tvar = year, gvar = first.treat, data = mpdta, vcov = ~countyreal,
+  cgroup = "never" # <= use never treated as control group
+  ) |>
+  emfx(type = "event")
+
+plot(mod_es2)
+
+## -------------------------------------------------------------------------------------------------
+plot(
+  mod_es2,
+  type = "ribbon",
+  col  = "darkcyan",
+  xlab = "Years post treatment",
+  main = "Minimum wage effect on (log) teen employment",
+  sub  = "Note: Using never-treated as control group",
+  # file = "event-study.png", width = 8, height = 5. ## uncomment to save file
+)
+
+## ----warning=FALSE, message=FALSE-----------------------------------------------------------------
+# install.packages("ggplot2")
+library(ggplot2)
+theme_set(
+  theme_linedraw() + theme(panel.grid = element_blank())
+)
+ggplot(mod_es2, aes(x = event, y = estimate, ymin = conf.low, ymax = conf.high)) +
+  geom_hline(yintercept = 0, lty = 2, col = "grey50") +
+  geom_vline(xintercept = -1, lty = 2, col = "grey50") +
+  geom_pointrange(col = "darkcyan") +
+  labs(
+    x = "Years post treatment",
+    y = "ATT",
+    title = "Effect on log teen employment",
+    caption = "Note: Using never-treated as control group"
+  )
+
+## -------------------------------------------------------------------------------------------------
+# install.packages("modelsummary")
 library(modelsummary)
 
 # Quick renaming function to replace ".Dtreat" with something more meaningful
@@ -45,67 +87,53 @@ rename_fn = function(old_names) {
 }
 
 modelsummary(
-  mod_es,
+  list(mod_es2, mod_es),
   shape       = term:event:statistic ~ model,
   coef_rename = rename_fn,
   gof_omit    = "Adj|Within|IC|RMSE",
+  stars       = TRUE,
   title       = "Event study",
   notes       = "Std. errors are clustered at the county level"
 )
 
-## -----------------------------------------------------------------------------
-library(ggplot2)
-theme_set(
-  theme_minimal() + theme(panel.grid.minor = element_blank())
-)
-
-ggplot(mod_es, aes(x = event, y = estimate, ymin = conf.low, ymax = conf.high)) +
-  geom_hline(yintercept = 0) +
-  geom_pointrange(col = "darkcyan") +
-  labs(x = "Years post treatment", y = "Effect on log teen employment")
-
-## -----------------------------------------------------------------------------
-# Use post_only = FALSE to get the "zero" pre-treatment effects
-mod_es2 = emfx(mod, type = "event", post_only = FALSE)
-
-ggplot(mod_es2, aes(x = event, y = estimate, ymin = conf.low, ymax = conf.high)) +
-  geom_hline(yintercept = 0) +
-  geom_vline(xintercept = -1, lty = 2) +
-  geom_pointrange(col = "darkcyan") +
-  labs(
-    x = "Years post treatment", y = "Effect on log teen employment",
-    caption = "Note: Zero pre-treatment effects for illustrative purposes only."
-  )
-
-## -----------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
 gls_fips = c("IL" = 17, "IN" = 18, "MI" = 26, "MN" = 27,
              "NY" = 36, "OH" = 39, "PA" = 42, "WI" = 55)
 
 mpdta$gls = substr(mpdta$countyreal, 1, 2) %in% gls_fips
 
-## -----------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
 hmod = etwfe(
-   lemp ~ lpop, tvar = year, gvar = first.treat, data = mpdta, 
-   vcov = ~countyreal,
-   xvar = gls           ## <= het. TEs by gls
-   )
-
-# Heterogeneous ATTs (could also specify `type = "event"`, etc.) 
-emfx(hmod)
-
-## -----------------------------------------------------------------------------
-emfx(hmod, hypothesis = "b1 = b2")
-
-## -----------------------------------------------------------------------------
-modelsummary(
-    models      = list("GLS county" = emfx(hmod)),
-    shape       = term + statistic ~ model + gls, # add xvar variable (here: gls)
-    coef_map    = c(".Dtreat" = "ATT"),
-    gof_map     = NA,
-    title       = "Comparing the ATT on GLS and non-GLS counties"
+  lemp ~ lpop, tvar = year, gvar = first.treat, data = mpdta,
+  cgroup = "never",
+  vcov = ~countyreal,
+  xvar = gls           ## <= het. TEs by gls
 )
 
-## ----warning=FALSE, message=FALSE---------------------------------------------
+# Simple heterogeneous ATTs (could also specify `type = "event"`, etc.) 
+(hmod_mfx = emfx(hmod))
+
+## ----warning=FALSE--------------------------------------------------------------------------------
+emfx(hmod, hypothesis = "b1 = b2")
+
+## -------------------------------------------------------------------------------------------------
+plot(hmod_mfx)
+
+## -------------------------------------------------------------------------------------------------
+hmod |>
+  emfx("event") |>
+  plot(type = "ribbon")
+
+## -------------------------------------------------------------------------------------------------
+modelsummary(
+  models      = list("GLS county" = hmod_mfx),
+  shape       = term + statistic ~ model + gls, # add xvar variable (here: gls)
+  coef_map    = c(".Dtreat" = "ATT"),
+  gof_map     = NA,
+  title       = "Comparing the ATT on GLS and non-GLS counties"
+)
+
+## ----warning=FALSE, message=FALSE-----------------------------------------------------------------
 mpdta$emp = exp(mpdta$lemp)
 
 etwfe(
@@ -114,32 +142,32 @@ etwfe(
   ) |>
   emfx("event")
 
-## -----------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
 # Step 0 already complete: using the same `mod` object from earlier...
 
 # Step 1
 emfx(mod, type = "event", vcov = FALSE)
 
 # Step 2
-emfx(mod, type = "event", vcov = FALSE, collapse = TRUE)
+emfx(mod, type = "event", vcov = FALSE, compress = TRUE)
 
 # Step 3: Results from 1 and 2 are similar enough, so get approx. SEs
-mod_es2 = emfx(mod, type = "event", collapse = TRUE)
+mod_es_compressed = emfx(mod, type = "event", compress = TRUE)
 
-## -----------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
 modelsummary(
-    list("Original" = mod_es, "Collapsed" = mod_es2),
-    shape       = term:event:statistic ~ model,
-    coef_rename = rename_fn,
-    gof_omit    = "Adj|Within|IC|RMSE",
-    title       = "Event study",
-    notes       = "Std. errors are clustered at the county level"
+  list("Original" = mod_es, "Compressed" = mod_es_compressed),
+  shape       = term:event:statistic ~ model,
+  coef_rename = rename_fn,
+  gof_omit    = "Adj|Within|IC|RMSE",
+  title       = "Event study",
+  notes       = "Std. errors are clustered at the county level"
 )
 
-## -----------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
 mod$fml_all
 
-## -----------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
 # First construct the dataset
 mpdta2 = mpdta |>
   transform(
@@ -155,13 +183,13 @@ mod2 = fixest::feols(
   vcov = ~countyreal
 )
 
-## -----------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
 modelsummary(
   list("etwfe" = mod, "manual" = mod2),
   gof_map = NA # drop all goodness-of-fit info for brevity
 )
 
-## -----------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
 library(marginaleffects)
 
 # Simple ATT
@@ -170,7 +198,7 @@ slopes(
   newdata   = subset(mpdta2, .Dtreat), # we only want rows where .Dtreat is TRUE
   variables = ".Dtreat", 
   by        = ".Dtreat"
-  )
+)
 
 # Event study
 slopes(
@@ -178,9 +206,9 @@ slopes(
   newdata   = transform(subset(mpdta2, .Dtreat), event = year - first.treat),
   variables = ".Dtreat", 
   by        = "event"
-  )
+)
 
-## ----message=FALSE, warning=FALSE---------------------------------------------
+## ----message=FALSE, warning=FALSE-----------------------------------------------------------------
 # fe = "feo" (fixed effects only)
 mod_feo = etwfe(
   lemp ~ lpop, tvar = year, gvar = first.treat, data = mpdta, vcov = ~countyreal,
@@ -207,7 +235,7 @@ mod_none2 = fixest::feols(
   data = mpdta2, vcov = ~countyreal
 )
 
-## -----------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
 mods = list(
   "etwfe"         = mod,
   "manual"        = mod2,
@@ -219,7 +247,7 @@ mods = list(
 
 modelsummary(mods, gof_map = NA)
 
-## -----------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
 mod_es_i = etwfe(
   lemp ~ lpop, tvar = year, gvar = first.treat, data = mpdta,
   ivar = countyreal  # NEW: Use unit-level (county) FEs